+++ ariles/include/ariles2/visitors/defaults.h
@@ -109,7 +109,7 @@ namespace ariles2
 #define ARILES2_NAMED_ENTRY_defaults(v, entry, name) visitor.visitMapEntry(entry, #name, parameters);
 #define ARILES2_PARENT_defaults(v, entry)
 #define ARILES2_VISIT_defaults                                                                                         \
-    template <class t_Visitor>                                                                                         \
+    template <class t_Visitor> /* cppcheck-suppress duplInheritedMember */                                             \
     void arilesVisit(                                                                                                  \
             const t_Visitor &visitor,                                                                                  \
             const typename t_Visitor::Parameters &parameters,                                                          \