+++ ariles/include/ariles2/visitors/finalize.h
@@ -48,7 +48,7 @@ namespace ariles2
 #define ARILES2_NAMED_ENTRY_finalize(v, entry, name) visitor.visitMapEntry(entry, #name, parameters);
 #define ARILES2_PARENT_finalize(v, entry)
 #define ARILES2_VISIT_finalize                                                                                         \
-    template <class t_Visitor>                                                                                         \
+    template <class t_Visitor> /* cppcheck-suppress duplInheritedMember */                                             \
     void arilesVisit(                                                                                                  \
             const t_Visitor &visitor,                                                                                  \
             const typename t_Visitor::Parameters &parameters,                                                          \