+++ ariles/tests/api_v2/demo_api_v2.cpp
@@ -7,6 +7,7 @@
 
     @brief
 */
+// cppcheck-suppress-file duplInheritedMember
 
 
 // ============================================================================