+++ ariles/tests/api_v2/types/minimal.h
@@ -6,6 +6,7 @@
 
     @brief
 */
+// cppcheck-suppress-file duplInheritedMember
 
 #pragma once