+++ ariles/tests/api_v2/types/pointers.h
@@ -5,6 +5,7 @@
 
     @brief
 */
+// cppcheck-suppress-file duplInheritedMember
 
 #pragma once